+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-
+* Use strand<> rather than legacy executor io_context::strand.
 * Use dispatch/post free functions to be independent of executor concept.
 * New name for polymorphic executor. Trait for detecting new executors.
 * Handler invoke and allocation hooks are deprecated.